requirements.txt
# Runtime dependency for `scripts/ai_backend.py`
openai~=1.86
cbor2~=5.6
pytest-cov~=6.2
cryptography~=45.0
keyring~=25.6
flask~=3.1
PyYAML~=6.0
psutil~=7.0
fastapi~=0.115
docker~=7.1
PyJWT~=2.10
redis~=6.2
rq~=1.16
uvicorn~=0.34
websockets~=12.0
pygit2~=1.15
argcomplete~=3.3
